gemm block quantization infrastructure for llm decoder style (#6814)
* another mha oob read

+++ src/layer/vulkan/multiheadattention_vulkan.cpp
@@ -48,6 +48,12 @@ int MultiHeadAttention_vulkan::load_param(const ParamDict& pd)
 
 int MultiHeadAttention_vulkan::create_pipeline(const Option& opt)
 {
+    if (weight_block_quantize)
+    {
+        NCNN_LOGE("MultiHeadAttention weight block quantization is not supported by Vulkan");
+        return -1;
+    }
+
     // const int embed_dim_per_head = embed_dim / num_heads;
     const int qdim = weight_data_size / embed_dim;
     {
@@ -255,6 +261,9 @@ int MultiHeadAttention_vulkan::create_pipeline(const Option& opt)
 
 int MultiHeadAttention_vulkan::destroy_pipeline(const Option& opt)
 {
+    if (weight_block_quantize)
+        return 0;
+
     if (q_gemm)
     {
         q_gemm->destroy_pipeline(opt);
@@ -326,6 +335,12 @@ int MultiHeadAttention_vulkan::destroy_pipeline(const Option& opt)
 
 int MultiHeadAttention_vulkan::upload_model(VkTransfer& cmd, const Option& opt)
 {
+    if (weight_block_quantize)
+    {
+        NCNN_LOGE("MultiHeadAttention weight block quantization is not supported by Vulkan");
+        return -1;
+    }
+
     if (q_gemm)
     {
         q_gemm->upload_model(cmd, opt);
@@ -351,6 +366,12 @@ int MultiHeadAttention_vulkan::upload_model(VkTransfer& cmd, const Option& opt)
 
 int MultiHeadAttention_vulkan::forward(const std::vector<VkMat>& bottom_blobs, std::vector<VkMat>& top_blobs, VkCompute& cmd, const Option& opt) const
 {
+    if (weight_block_quantize)
+    {
+        NCNN_LOGE("MultiHeadAttention weight block quantization is not supported by Vulkan");
+        return -1;
+    }
+
     int q_blob_i = 0;
     int k_blob_i = 0;
     int v_blob_i = 0;
